+++ electron/app/components/Samples.tsx
@@ -1,14 +1,14 @@
 import React, { useState, useEffect } from "react";
 import InfiniteScroll from "react-infinite-scroller";
-import { useSetRecoilState } from "recoil";
+import { useSetRecoilState, useRecoilValue } from "recoil";
 import { Grid, Loader, Dimmer } from "semantic-ui-react";
 import Sample from "./Sample";
 import connect from "../utils/connect";
 import tile from "./Samples.hooks";
 import * as atoms from "../recoil/atoms";
 
-function Samples(props) {
-  const { displayProps, state, setView, port } = props;
+function Samples({ setView }) {
+  const port = useRecoilValue(atoms.port);
   const setCurrentSamples = useSetRecoilState(atoms.currentSamples);
 
   const [scrollState, setScrollState] = tile(port);
@@ -32,11 +32,7 @@ function Samples(props) {
         <Grid columns={r.samples.length} style={r.style} key={i}>
           {r.samples.map((s, j) => (
             <Grid.Column key={j} style={{ padding: 0, width: "100%" }}>
-              <Sample
-                displayProps={displayProps}
-                sample={s}
-                setView={setView}
-              />
+              <Sample sample={s} setView={setView} />
             </Grid.Column>
           ))}
         </Grid>
@@ -45,4 +41,4 @@ function Samples(props) {
   );
 }
 
-export default connect(Samples);
+export default Samples;