+++ electron/app/recoil/atoms.ts
@@ -52,7 +52,7 @@ export const filterIncludeLabels = atomFamily({
 
 export const filterLabelConfidenceRange = atomFamily({
   key: "filterLabelConfidenceRange",
-  default: [0, 1],
+  default: [null, null],
 });
 
 export const filterLabelIncludeNoConfidence = atomFamily({
@@ -67,7 +67,7 @@ export const modalFilterIncludeLabels = atomFamily({
 
 export const modalFilterLabelConfidenceRange = atomFamily({
   key: "modalFilterLabelConfidenceRange",
-  default: [0, 1],
+  default: [null, null],
 });
 
 export const modalFilterLabelIncludeNoConfidence = atomFamily({
@@ -103,4 +103,14 @@ export const activeTags = atom({
 export const modalActiveTags = atom({
   key: "modalActiveTags",
   default: {},
+});
+
+export const filterNumericFieldRange = atomFamily({
+  key: "filterNumericFieldRange",
+  default: [null, null],
+});
+
+export const filterNumericFieldIncludeNone = atomFamily({
+  key: "filterNumericFieldIncludeNone",
+  default: true,
 });