+++ tests/utils/pytest_wrapper.py
@@ -8,10 +8,12 @@ import psutil
 import pytest
 
 try:
-    pytest.main(sys.argv[1:])
+    code = pytest.main(sys.argv[1:])
 finally:
     for child in psutil.Process().children(recursive=True):
         try:
             child.kill()
         except psutil.Error:
             pass
+
+exit(code)